# apps/tapcap-supplier/ClientCS/src/TapCap.Supplier/Model/TapCapBrokerPurchase.cs
/* 
 * THE TapCap supply
 *
 * The interace for TapCap between buyers & sellers.
 *
 * OpenAPI spec version: 0.1.0
 * Contact: stephen.taylor.dev@gmail.com
 * Generated by: https://github.com/openapitools/openapi-generator.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;
using OpenAPIDateConverter = TapCap.Supplier.Client.OpenAPIDateConverter;

namespace TapCap.Supplier.Model
{
    /// <summary>
    /// TapCapBrokerPurchase
    /// </summary>
    [DataContract]
    public partial class TapCapBrokerPurchase :  IEquatable<TapCapBrokerPurchase>
    {
        /// <summary>
        /// Initializes a new instance of the <see cref="TapCapBrokerPurchase" /> class.
        /// </summary>
        [JsonConstructorAttribute]
        protected TapCapBrokerPurchase() { }
        /// <summary>
        /// Initializes a new instance of the <see cref="TapCapBrokerPurchase" /> class.
        /// </summary>
        /// <param name="signedRequest">signedRequest (required).</param>
        /// <param name="fxRate">fxRate (required).</param>
        /// <param name="coinCharge">coinCharge (required).</param>
        /// <param name="cryptoCertificate">cryptoCertificate (required).</param>
        public TapCapBrokerPurchase(SignedMessage signedRequest = default(SignedMessage), FXRate fxRate = default(FXRate), long? coinCharge = default(long?), byte[] cryptoCertificate = default(byte[]))
        {
            // to ensure "signedRequest" is required (not null)
            if (signedRequest == null)
            {
                throw new InvalidDataException("signedRequest is a required property for TapCapBrokerPurchase and cannot be null");
            }
            else
            {
                this.SignedRequest = signedRequest;
            }
            // to ensure "fxRate" is required (not null)
            if (fxRate == null)
            {
                throw new InvalidDataException("fxRate is a required property for TapCapBrokerPurchase and cannot be null");
            }
            else
            {
                this.FxRate = fxRate;
            }
            // to ensure "coinCharge" is required (not null)
            if (coinCharge == null)
            {
                throw new InvalidDataException("coinCharge is a required property for TapCapBrokerPurchase and cannot be null");
            }
            else
            {
                this.CoinCharge = coinCharge;
            }
            // to ensure "cryptoCertificate" is required (not null)
            if (cryptoCertificate == null)
            {
                throw new InvalidDataException("cryptoCertificate is a required property for TapCapBrokerPurchase and cannot be null");
            }
            else
            {
                this.CryptoCertificate = cryptoCertificate;
            }
        }
        
        /// <summary>
        /// Gets or Sets SignedRequest
        /// </summary>
        [DataMember(Name="signedRequest", EmitDefaultValue=false)]
        public SignedMessage SignedRequest { get; set; }

        /// <summary>
        /// Gets or Sets FxRate
        /// </summary>
        [DataMember(Name="FxRate", EmitDefaultValue=false)]
        public FXRate FxRate { get; set; }

        /// <summary>
        /// Gets or Sets CoinCharge
        /// </summary>
        [DataMember(Name="coinCharge", EmitDefaultValue=false)]
        public long? CoinCharge { get; set; }

        /// <summary>
        /// Gets or Sets CryptoCertificate
        /// </summary>
        [DataMember(Name="cryptoCertificate", EmitDefaultValue=false)]
        public byte[] CryptoCertificate { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class TapCapBrokerPurchase {\n");
            sb.Append("  SignedRequest: ").Append(SignedRequest).Append("\n");
            sb.Append("  FxRate: ").Append(FxRate).Append("\n");
            sb.Append("  CoinCharge: ").Append(CoinCharge).Append("\n");
            sb.Append("  CryptoCertificate: ").Append(CryptoCertificate).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }
  
        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public virtual string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="input">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object input)
        {
            return this.Equals(input as TapCapBrokerPurchase);
        }

        /// <summary>
        /// Returns true if TapCapBrokerPurchase instances are equal
        /// </summary>
        /// <param name="input">Instance of TapCapBrokerPurchase to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(TapCapBrokerPurchase input)
        {
            if (input == null)
                return false;

            return 
                (
                    this.SignedRequest == input.SignedRequest ||
                    (this.SignedRequest != null &&
                    this.SignedRequest.Equals(input.SignedRequest))
                ) && 
                (
                    this.FxRate == input.FxRate ||
                    (this.FxRate != null &&
                    this.FxRate.Equals(input.FxRate))
                ) && 
                (
                    this.CoinCharge == input.CoinCharge ||
                    (this.CoinCharge != null &&
                    this.CoinCharge.Equals(input.CoinCharge))
                ) && 
                (
                    this.CryptoCertificate == input.CryptoCertificate ||
                    (this.CryptoCertificate != null &&
                    this.CryptoCertificate.Equals(input.CryptoCertificate))
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                int hashCode = 41;
                if (this.SignedRequest != null)
                    hashCode = hashCode * 59 + this.SignedRequest.GetHashCode();
                if (this.FxRate != null)
                    hashCode = hashCode * 59 + this.FxRate.GetHashCode();
                if (this.CoinCharge != null)
                    hashCode = hashCode * 59 + this.CoinCharge.GetHashCode();
                if (this.CryptoCertificate != null)
                    hashCode = hashCode * 59 + this.CryptoCertificate.GetHashCode();
                return hashCode;
            }
        }
    }

}